:sparkles: Se agrega rutas de products

+++ backend/src/models/product.rs
@@ -8,8 +8,8 @@ pub struct Product {
     pub name: String,
     pub image: String,
     pub price: f64,
-    pub units: i32,
-    pub deleted: i8, // 0 false | 1 true
+    pub units: i64,
+    pub deleted: i64, // 0 false | 1 true
 }
 
 #[derive(Debug, FromRow, Serialize, Deserialize)]
@@ -18,6 +18,6 @@ pub struct CreateProduct {
     pub name: String,
     pub image: String,
     pub price: f64,
-    pub units: i32,
-    pub deleted: i8, // 0 false | 1 true
+    pub units: i64,
+    pub deleted: i64, // 0 false | 1 true
 }
